telegram_bot_docker/telegram_bot_docker/requirements.txt
python-telegram-bot>=20,<22

